优化

+++ src/backend.py
@@ -1,8 +1,11 @@
+import threading
+
 import requests
 from flask import Flask, jsonify, request
 from ftplib import FTP
 from config import ftp_config, mqtt_config, backend_config
 from flask_cors import CORS, cross_origin
+import time
 
 import paho.mqtt.client as mqtt
 import os
@@ -25,14 +28,29 @@ def get_data():
         try:
             file_name = message
             file_path = os.path.join('../public/pdf/', file_name)
-            with open(file_path, 'wb') as f:
-                ftp.retrbinary('RETR ' + message, f.write)
-            print(f'下载{message}成功！')
-            return jsonify({'status': '成功', 'path': file_path})
-        except Exception as e:
-            print(f'下载{message}文件失败：{str(e)}')
-            return jsonify({'status': '失败'})
 
+            if os.path.isfile(file_path):
+                print(f'已使用本地文件{file_path}！')
+                return jsonify({'status': '本地文件', 'path': file_path})
+            else:
+                with open(file_path, 'wb') as f:
+                    ftp.retrbinary('RETR ' + message, f.write)
+                print(f'下载{message}成功！')
+                return jsonify({'status': '成功', 'path': file_path})
+        except Exception as err:
+            print(f'下载{message}文件失败：{str(err)}')
+            if os.path.isfile(file_path):
+                print(f'已使用本地文件{file_path}！')
+                return jsonify({'status': '本地文件', 'path': file_path})
+            else:
+                print(f'本地文件{file_path}不存在！')
+                return jsonify(({'status': '失败'}))
+
+
+@app.route('/api/warning', methods=['POST'])
+@cross_origin(origin='http://localhost:5173', supports_credentials=True)
+def get_warning():
+    return {'isWarning': isWarning}
 
 
 def setup_ftp():  # 连接ftp服务器并下载json文件
@@ -55,15 +73,31 @@ def setup_ftp():  # 连接ftp服务器并下载json文件
         return
 
 
-# 用于接收到消息时的回调函数
+isWarning = False
+lastTimeReceived = None
+
+
+# 接受报警信息、设置报警标志位
 def on_message(client, userdata, msg):
-    print("Received message: " + msg.payload.decode())
+    global isWarning, lastTimeReceived
+    isWarning = True
+    lastTimeReceived = time.time()
+    print("收到报警信息！")
+
+
+# 每5s检查一次报警信息，如果没有报警，则设置报警标志位
+def check_warning():
+    global isWarning, lastTimeReceived
+    if isWarning and time.time() - lastTimeReceived > 5:
+        isWarning = False
+        print("报警信息消失！")
+    threading.Timer(5, check_warning).start()
 
 
 def setup_mqtt():
     try:
         client = mqtt.Client()
-        client.connect(mqtt_config['url'], 1883, 60)
+        client.connect(mqtt_config['url'], mqtt_config['port'], 60)
         print('连接mqtt服务器成功！')
     except Exception as e:
         print(f'连接mqtt服务器失败：{str(e)}')
@@ -96,6 +130,7 @@ def download_ftp(file_name):
 
 setup_ftp()
 setup_mqtt()
+check_warning()
 
 if __name__ == '__main__':
     # app.run(port=backend_config['port'])
